refactor: update import paths for attachment authorization classes
